Move example commands into examples/RoboFile.php, and commands used in tests into tests/src/TestRoboFile.php.
